refactor: instance
